Restructure parser interface
Bring out a more complete CsvParser.Setup object, which can be used in
the StoreView/Inspect/Parser API pages to allow more control over
not-yet-parsed data.  In particular, the Parser page can view the
unparsed data and show how it would parse with or without the 1st row as
a header.  Currently this flag is not tied back into the parser, but at
least the data is now available.  Major cleanups of StoreView display of
unparsed data.