Survive TCP resets under heavy load
After doing a TCP open, also send the 1st ByteBuffer load... all inside a
"catch all & retry" loop.  If either the open fails or the write fails for
excessive system load, we'll catch the error, log it, stall for 1/2sec (to
allow congestion to settle down) and try again.  It's not perfect: fails
in the 2nd or later ByteBuffer loads will not be retried, but we appear to
only fail in the 1st load or not at all.