First cut writable fluid-vecs
API is set/set8 mirrors at/at8.
Compressed chunks support 2 new calls: writing into the compressed
chunk, or returning false if the value does not fit, and bulk expanding
the chunk into a NewChunk.
After all bulk writes are done, the chunk needs to be "closed" - which
will recompress NewChunks.
Any writing "opens" the Vec, triggers the wiping out of any cached
rollup data, such as min & max, and it is an error to ask for such data
until the entire Vec is "closed".  Longer term plan is to have MRTask2
do the "close" action (not implemented).