get s3 jar. Now you might ask, why is the internet in love with cats?
Among Unix fans, cat is considered an excellent example of user-interface design, because it delivers the file contents without such verbosity as spacing or headers between the files, and because it does not require the files to consist of lines of text, but works with any sort of data.

Among Unix critics, cat is considered the canonical example of bad user-interface design, because of its woefully unobvious name. It is far more often used to blast a single file to standard output than to concatenate two or more files. The name cat for the former operation is just as unintuitive as, say, LISP's cdr.

It is reasonable to say, this note should get a "Useless use of cat" award. (for pointing out the internet is just an implementation of 'cat')