Overhaul the memory access of GBM hot code
All rows, all cols, accumulate histograms.  This is the hot hot inner
loop of GBM, so we do some non-standard optimizations.  The rows in this
chunk are spread out amongst a modest set of NodeIDs/splits.  Normally
we would visit the rows in row-order, but this visits the NIDs in random
order.  The hot-part of this code updates the histograms racily (via
atomic updates) - once-per-row.  This optimized version updates the
histograms once-per-NID, but requires pre-sorting the rows by NID.
Speed on bigger data is 10x or 20x faster.
nnids[] array holds node-relative indices (_leaf pre-subtracted).