Exec2: more tolerance of operators as values.
As of now, operators and functions can be assigned and used as if
normal values. Attention should be paid to operators with arity
ambiguity, such as + and -. In parsing of an infix operator, the
the arity is inferred by how the operator is placed in the
expression. In parsing of an value, the binary form of the operator is
used.

Some examples

x<-+;x(1,2) // returns 3

x<-+;x(2) // error, expects two arguments

-(1)      // returns -1

(-)(1)    // error, expects two arguments

(0?+:-)(1,2) // returns 3