HEX-1573: s3n load fails on S3Exception.
Load from S3 via S3N URL via HDFS client failed on S3Exception.
The problem was tricky.

We are supporting different HDFS versions.
New version declares S3Exception as IOException (so it was properly
catched).
But old versions (0.20.xxx) declares it as RuntimeException and S3N load
was failing because of the uncatched exception if S3 errors appeared.