Fix for prediction on large trees.
Missing conversion to float caused differences between interpreted
and generated version of trees.

Moreover, I there are now 2 GBM tests generating 5000 trees and
verifying that the generated model can be compiled.
However, there are not passing now because on a small difference
on 7th decimal digit which needs to be tracked.