Changed locking operations to preserve _persist flag.
Default TAtomic behavior is to set _persist to ICED and not on disk, which is correct for most atomic operations which update the data. However, locking does not update the data itself (only changes transient members) and it was causing problems when locking lazy chunked(disk mapped) values backed by files on nfs/hdfs/s3. The lazy chunking would not work after locking.