Comment out Confusion Matrix-related checks (waiting for PUB-216 fix).
Add check for labels in addition to scores.

Add fast_mode back to NN and Neurons (which is set to false in order to pass
the check against the reference).  Fast mode can be over 2x faster, without
much change in accuracy, so it is enabled by default.  It skips backpropagation
from neurons with 0 activity, even if there could have been a change to
incoming weights due to L1/L2/momentum.