Tree model contains keys instead of compressed models.
PUB-308: the problem was that a tree model became to large
to be stored in DKV because of generated huge trees.

Hence, the fix saves each individual tree as
an individual object into DKV and model holds only a list of trees' keys.
On the other hand, a tree model has to make DKV#get to obtain a tree, hence it preserve
a tree cache for already loaded trees.