support type 7 interpolation between two values (default). matches R default and numpy.percentile default for numpy v1.8. numpy.percentile v1.9 will add choices so you can get both cases.... scipy mquantiles doesn't default to type 7 but type 7 behavior can be forced with params. Type 2 is mean between two values.
The in-between stuff is only when the threshold lands in between rows in the dataset..it's not an 'estimate' for the dataset, it's an exact answer using two adjacent values in the virtual sorted list. ..seems to match comparing to np.percentile and scipy mquantiles with alphab=1 and betab=b..i.e. the binning needs to give you two results (in some cases) that you then use to create the answer.