Fix of PUB-586
The problem: R does not send a boolean flag for classification.
H2O server part by default expects classification to be true
and rejects all requests which try to do classification on float-column.
However, classification UI control is not invoked (!!!) if classification
flag is not specified (since it is not compulsory).
As a result,

Possible solutions:
 (1) make classification flag compulsory. In this case, R has to always
   sent it, and R has to decided if user requests classification or
regression.
 (2) have 3-state boolean field for classification - not-specified, specified-true,
   specified-false
  - in this case, if classificaiton is not-specified, H2O server side
    can decide based on data which strategy (classification/regression)
should be used
 (3) switch flag on server side. In this case, even if classification
flag is true server decide to switch to regression if the response
column is float

Selected solution: (3) (to deliver quick patch), but
it would be nice to inform user about decision that regression
was selected.