PUB-923: Fix K-Means to pass golden R test.
Problem: within_cluster_variances was computed correctly, but the name was
wrong, since the R wrapper equates this value with withinss, which is the
within_cluster_sum_of_squares. So in order to not upset the R side of things,
we compute sum of squares instead of variances.

Also re-enable returning total_within_SS from R.