Fixes for test_create_frame_rand1.py
1. Add more bounds checking for _mins.length 0 and _maxs.length 0.
Was getting an AIOOBE for synthetic frame from CreateFrame.

2. Disable some assertion checks when there is only one bucket.
GET  /2/CreateFrame.json randomize=1 integer_fraction=0 real_range=1234567890 response_factors=0 factors=1
was able to create a frame that tripped this assertion.

This may or may not be OK.

The vec thought it was isEnum() true, but the Chunk had negative double values (C8DChunk).